RF - moved direction getters to dipy/direction
