+++ dipy/direction/__init__.py
@@ -1,4 +1,5 @@
 
-from .probabilistic_direction_getter import ProbabilisticDirectionGetter
-from .probabilistic_direction_getter import DeterministicMaximumDirectionGetter
+from .pmf_direction_getter import ProbabilisticDirectionGetter
+from .pmf_direction_getter import DeterministicMaximumDirectionGetter
+from .direction_getter import DirectionGetter
 from .peaks import *
