+++ dipy/direction/direction_getter.pxd
