+++ dipy/direction/pmf_direction_getter.py
@@ -4,7 +4,7 @@ discrete distribution (pmf) at each step of the tracking."""
 import numpy as np
 from dipy.direction.peaks import peak_directions, default_sphere
 from dipy.reconst.shm import order_from_ncoef, sph_harm_lookup
-from dipy.tracking.local.direction_getter import DirectionGetter
+from dipy.direction.direction_getter import DirectionGetter
 from dipy.tracking.local.interpolation import trilinear_interpolate4d
