+++ dipy/tracking/local/__init__.py
@@ -1,7 +1,6 @@
 from .localtracking import LocalTracking
 from .tissue_classifier import (ActTissueClassifier, BinaryTissueClassifier,
                                 ThresholdTissueClassifier, TissueClassifier)
-from .direction_getter import DirectionGetter
 from dipy.tracking import utils
 
 __all__ = ["ActTissueClassifier", "BinaryTissueClassifier", "LocalTracking",
